Facebook intergration and Twitter double intergration
So I have two options for twitter and one for facebook. The twitter
options are the embeded timeline (see actual footer) or, I finally got
the new twitter api 1.1 working with the OAuth. So I have a raw JSON
feed of the most resent tweets we can style however we like.